recipe/meta.yaml: change SHA256 digest (?!)
As detailed in a comment in the meta.yaml file, conda-build finds a different
value for the file checksum than I do manually. But I get the same value in an
anvil container on my laptop as the CI systems do, so at least its weirdness
seems to be self-consistent.