Switch non-Windows builds to CMake too; drop Fortran dep
Version 4.5.0 switches the build system to fully embrace the autotools
and libtool. If we need to change the build system, we might as well
jump to CMake, which is what the Windows builds are already using.

The CMake build doesn't actually do anything with a Fortran compiler,
and upon investigation the Autotools build hardly does either. The only
Fortran source files are demos, and the Fortran support in the C library
can be compiled without actually using the Fortran compiler; the Fortran
compiler is just used to figure out what kind of name mangling to use.
This scheme appears to default sensibly, and there's also the fact that
there probably aren't any packages in conda-forge that are interfacing
with CFITSTIO *in Fortran*. So, let's just drop the dependency on the
Fortran compiler.